Moved the tests that rely on permission bits to a new test file. The permission bits are not executed during windows builds.
